backend/requirements.txt
Django==5.0.1
Pillow==10.2.0
django-allauth==0.60.1
djangorestframework==3.14.0
django-rest-swagger==2.2.0
drf_yasg==1.21.7
django-filter==23.5
python-dotenv==1.0.1
setuptools==69.0.3
icecream==2.1.3
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
telebot==0.0.5
aiohttp==3.9.3
yookassa==3.0.1
django-colorfield
psycopg2-binary
django-localflavor==4.0
django-ckeditor-5==0.2.12
django-image-uploader-widget==0.4.4